+++ web/src/lib/components/assets/thumbnail/image-thumbnail.svelte
@@ -1,13 +1,10 @@
 <script lang="ts">
-  import { thumbhash } from '$lib/actions/thumbhash';
   import BrokenAsset from '$lib/components/assets/broken-asset.svelte';
   import Icon from '$lib/components/elements/icon.svelte';
   import { cancelImageUrl } from '$lib/utils/sw-messaging';
-  import { TUNABLES } from '$lib/utils/tunables';
   import { mdiEyeOffOutline } from '@mdi/js';
   import type { ActionReturn } from 'svelte/action';
   import type { ClassValue } from 'svelte/elements';
-  import { fade } from 'svelte/transition';
 
   interface Props {
     url: string;
@@ -15,7 +12,6 @@
     title?: string | null;
     heightStyle?: string | undefined;
     widthStyle: string;
-    base64ThumbHash?: string | null;
     curve?: boolean;
     shadow?: boolean;
     circle?: boolean;
@@ -33,7 +29,6 @@
     title = null,
     heightStyle = undefined,
     widthStyle,
-    base64ThumbHash = null,
     curve = false,
     shadow = false,
     circle = false,
@@ -44,10 +39,6 @@
     class: imageClass = '',
     brokenAssetClass = '',
   }: Props = $props();
-
-  let {
-    IMAGE_THUMBNAIL: { THUMBHASH_FADE_DURATION },
-  } = TUNABLES;
 
   let loaded = $state(false);
   let errored = $state(false);
@@ -100,7 +91,6 @@
     alt={loaded || errored ? altText : ''}
     {title}
     class={['object-cover', optionalClasses, imageClass]}
-    class:opacity-0={!thumbhash && !loaded}
     draggable="false"
   />
 {/if}
@@ -109,20 +99,4 @@
   <div class="absolute start-1/2 top-1/2 translate-x-[-50%] translate-y-[-50%] transform">
     <Icon {title} path={mdiEyeOffOutline} size="2em" class={hiddenIconClass} />
   </div>
-{/if}
-
-{#if base64ThumbHash && (!loaded || errored)}
-  <canvas
-    use:thumbhash={{ base64ThumbHash }}
-    data-testid="thumbhash"
-    style:width={widthStyle}
-    style:height={heightStyle}
-    {title}
-    class="absolute top-0 object-cover"
-    class:rounded-xl={curve}
-    class:shadow-lg={shadow}
-    class:rounded-full={circle}
-    draggable="false"
-    out:fade={{ duration: THUMBHASH_FADE_DURATION }}
-  ></canvas>
 {/if}
